kernel: use kvrealloc
We used to think that oneplus protected the original memory with RO, but we are wrong. Why kfree() panic kernel is that original memory isn't allocated with kmalloc() but kvmalloc(). Normal device will eventually use kmalloc() anyway, but oneplus seems to use more aggressive memory allocation optimization and switched to vmalloc() path.

Use kvrealloc() to fix it. Note, kvrealloc() will check flags first and use kmalloc() if GFP_ATOMIC is passed, so it is fine to be called in kprobe context. This also fix the last memory leak we found in ksu driver.